Refactor selection logic: move mergePolygons and mergePointSetsToHull functions to polygonUtils.js; add SVG drawing utilities for polygons and bounding boxes
