+++ html/scripts/growSelection.js
@@ -1,6 +1,7 @@
 import { showError } from './main.js';
 import { selectionSize } from './settings.js';
-import { getPathPoints, isPointInPolygon } from './polygonUtils.js';
+import { getPathPoints, isPointInPolygon, mergePolygons, mergePointSetsToHull } from './polygonUtils.js';
+import { drawPolygon, drawBBoxRect } from './svgDebugUtils.js';
 
 export function growSelection() {
     const mergedPolygon = getSelectionBoundingPolygon();
@@ -108,62 +109,6 @@ function getSelectionBoundingPolygon() {
     return mergedPolygon;
 }
 
-function mergePolygons(polygon1, polygon2) {
-    if (polygon1.length === 0) return polygon2;
-    if (polygon2.length === 0) return polygon1;
-    const allPoints = [...polygon1, ...polygon2];
-    return mergePointSetsToHull(allPoints);
-}
-
-function mergePointSetsToHull(points) {
-    points.sort((a, b) => a.x === b.x ? a.y - b.y : a.x - b.x);
-    const cross = (o, a, b) => (a.x - o.x) * (b.y - o.y) - (a.y - o.y) * (b.x - o.x);
-    const mergeHulls = (hull1, hull2) => {
-        const allPoints = [...hull1, ...hull2];
-        allPoints.sort((a, b) => a.x === b.x ? a.y - b.y : a.x - b.x);
-        const lower = [];
-        for (const point of allPoints) {
-            while (lower.length >= 2 && cross(lower[lower.length - 2], lower[lower.length - 1], point) <= 0) {
-                lower.pop();
-            }
-            lower.push(point);
-        }
-        const upper = [];
-        for (let i = allPoints.length - 1; i >= 0; i--) {
-            const point = allPoints[i];
-            while (upper.length >= 2 && cross(upper[upper.length - 2], upper[upper.length - 1], point) <= 0) {
-                upper.pop();
-            }
-            upper.push(point);
-        }
-        upper.pop();
-        lower.pop();
-        return lower.concat(upper);
-    };
-    let mergedHull = [];
-    for (let i = 0; i < points.length - 1; i++) {
-        const pair = [points[i], points[i + 1]];
-        const pairHull = mergeHulls(pair, []);
-        mergedHull = mergeHulls(mergedHull, pairHull);
-    }
-    return mergedHull;
-}
-
-function drawPolygon(polygon, color, id) {
-    if (!(location.hostname === 'localhost' || location.hostname === '127.0.0.1')) return;
-    let existingPolygon = document.querySelector(`#${id}`);
-    if (existingPolygon) existingPolygon.remove();
-    const svgElement = document.querySelector('#map-container svg');
-    const polygonElement = document.createElementNS('http://www.w3.org/2000/svg', 'polygon');
-    polygonElement.setAttribute('id', id);
-    polygonElement.setAttribute('points', polygon.map(p => `${p.x},${p.y}`).join(' '));
-    polygonElement.setAttribute('fill', color);
-    polygonElement.setAttribute('stroke', '#000');
-    polygonElement.setAttribute('stroke-width', '1');
-    polygonElement.setAttribute('opacity', '0.5');
-    svgElement.appendChild(polygonElement);
-}
-
 function createExpandedPolygonCollection(polygon, expansionDiameter = selectionSize) {
     const expandedPolygons = [];
     const expansionRadius = expansionDiameter * 2;
@@ -178,24 +123,6 @@ function createExpandedPolygonCollection(polygon, expansionDiameter = selectionS
         expandedPolygons.push(expandedPoints);
     });
     return expandedPolygons;
-}
-
-function drawBBoxRect(bbox, color, id) {
-    if (!(location.hostname === 'localhost' || location.hostname === '127.0.0.1')) return;
-    let existingRect = document.querySelector(`#${id}`);
-    if (existingRect) existingRect.remove();
-    const svgElement = document.querySelector('#map-container svg');
-    const rect = document.createElementNS('http://www.w3.org/2000/svg', 'rect');
-    rect.setAttribute('id', id);
-    rect.setAttribute('x', bbox.x);
-    rect.setAttribute('y', bbox.y);
-    rect.setAttribute('width', bbox.width);
-    rect.setAttribute('height', bbox.height);
-    rect.setAttribute('fill', 'none');
-    rect.setAttribute('stroke', color);
-    rect.setAttribute('stroke-width', '2');
-    rect.setAttribute('stroke-dasharray', '4,2');
-    svgElement.appendChild(rect);
 }
 
 // These must be provided by the main script
