+++ html/scripts/polygonUtils.js
@@ -48,3 +48,46 @@ export function isPathInSelection(path, polygon) {
     const { points } = getPathPoints(path);
     return points.some(point => isPointInPolygon(point, polygon)) || isPointInPolygon(points[0], polygon);
 }
+
+// Merge two polygons into a single convex hull
+export function mergePolygons(polygon1, polygon2) {
+    if (polygon1.length === 0) return polygon2;
+    if (polygon2.length === 0) return polygon1;
+    const allPoints = [...polygon1, ...polygon2];
+    return mergePointSetsToHull(allPoints);
+}
+
+// Compute the convex hull of a set of points
+export function mergePointSetsToHull(points) {
+    points.sort((a, b) => a.x === b.x ? a.y - b.y : a.x - b.x);
+    const cross = (o, a, b) => (a.x - o.x) * (b.y - o.y) - (a.y - o.y) * (b.x - o.x);
+    const mergeHulls = (hull1, hull2) => {
+        const allPoints = [...hull1, ...hull2];
+        allPoints.sort((a, b) => a.x === b.x ? a.y - b.y : a.x - b.x);
+        const lower = [];
+        for (const point of allPoints) {
+            while (lower.length >= 2 && cross(lower[lower.length - 2], lower[lower.length - 1], point) <= 0) {
+                lower.pop();
+            }
+            lower.push(point);
+        }
+        const upper = [];
+        for (let i = allPoints.length - 1; i >= 0; i--) {
+            const point = allPoints[i];
+            while (upper.length >= 2 && cross(upper[upper.length - 2], upper[upper.length - 1], point) <= 0) {
+                upper.pop();
+            }
+            upper.push(point);
+        }
+        upper.pop();
+        lower.pop();
+        return lower.concat(upper);
+    };
+    let mergedHull = [];
+    for (let i = 0; i < points.length - 1; i++) {
+        const pair = [points[i], points[i + 1]];
+        const pairHull = mergeHulls(pair, []);
+        mergedHull = mergeHulls(mergedHull, pairHull);
+    }
+    return mergedHull;
+}
